node backend

+++ js/item.js
@@ -1,13 +1,29 @@
 $(document).ready(function () {
-    const url = 'http://172.34.11.117:8000'
+    const url = 'http://localhost:4000'
+
+    const getToken = () => {
+        const token = sessionStorage.getItem('token');
+
+        if (!token) {
+            Swal.fire({
+                icon: 'warning',
+                text: 'You must be logged in to access this page.',
+                showConfirmButton: true
+            }).then(() => {
+                window.location.href = 'login.html';
+            });
+            return;
+        }
+        return JSON.parse(token)
+    }
 
     $('#itable').DataTable({
         ajax: {
             url: `${url}/api/v1/items`,
-            dataSrc: '',
-            // headers: {
-            //     "Authorization": "Bearer " + access_token 
-            // },
+            dataSrc: 'rows',
+            headers: {
+                "Authorization": "Bearer " + getToken()
+            },
         },
         dom: 'Bfrtip',
         buttons: [
@@ -36,7 +52,8 @@ $(document).ready(function () {
             { data: 'description' },
             { data: 'cost_price' },
             { data: 'sell_price' },
-            { data: 'stock.quantity' },
+            // { data: 'stock.quantity' },
+            { data: 'quantity' },
             {
                 data: null,
                 render: function (data, type, row) {
@@ -56,7 +73,7 @@ $(document).ready(function () {
         for (var pair of formData.entries()) {
             console.log(pair[0] + ', ' + pair[1]);
         }
-        // const token = getToken()
+        const token = getToken()
 
         $.ajax({
             method: "POST",
@@ -65,9 +82,9 @@ $(document).ready(function () {
             contentType: false,
             processData: false,
             dataType: "json",
-            // headers: {
-            //     "Authorization": "Bearer " + token
-            // },
+            headers: {
+                "Authorization": "Bearer " + token
+            },
             success: function (data) {
                 console.log(data);
                 $("#itemModal").modal("hide");
@@ -110,13 +127,13 @@ $(document).ready(function () {
             url: `${url}/api/v1/items/${id}`,
             dataType: "json",
             success: function (data) {
-                const { description, item_id, stock, cost_price, sell_price } = data
+                const { description, item_id, stock, cost_price, sell_price, quantity } = data.result[0]
 
-                console.log(data);
+                console.log(data.result[0]);
                 $('#desc').val(description)
                 $('#sell').val(sell_price)
                 $('#cost').val(cost_price)
-                $('#qty').val(stock.quantity)
+                $('#qty').val(quantity)
                 $("#iform").append(`<img src="${url}/${data.img_path}" width='200px', height='200px' id="itemImage"   />`)
 
             },
@@ -134,9 +151,10 @@ $(document).ready(function () {
 
         var data = $('#iform')[0];
         let formData = new FormData(data);
-        formData.append("_method", "PUT")
+        // formData.append("_method", "PUT")
         $.ajax({
-            method: "POST",
+            // method: "POST",
+            method: "PUT",
             url: `${url}/api/v1/items/${id}`,
             data: formData,
             contentType: false,
@@ -161,48 +179,49 @@ $(document).ready(function () {
         var id = $(this).data('id');
         var $row = $(this).closest('tr');
         console.log(id);
-        // if (getToken()) {
-        bootbox.confirm({
-            message: "do you want to delete this item",
-            buttons: {
-                confirm: {
-                    label: 'yes',
-                    className: 'btn-success'
+        if (getToken()) {
+            bootbox.confirm({
+                message: "do you want to delete this item",
+                buttons: {
+                    confirm: {
+                        label: 'yes',
+                        className: 'btn-success'
+                    },
+                    cancel: {
+                        label: 'no',
+                        className: 'btn-danger'
+                    }
                 },
-                cancel: {
-                    label: 'no',
-                    className: 'btn-danger'
-                }
-            },
-            callback: function (result) {
-                console.log(result);
-                if (result) {
-                    $.ajax({
-                        method: "DELETE",
-                        url: `${url}/api/v1/items/${id}`,
-                        dataType: "json",
-                        // headers: {
-                        //     "Authorization": "Bearer " + getToken()
-                        // },
-                        success: function (data) {
-                            console.log(data);
-                            $row.fadeOut(4000, function () {
-                                table.row($row).remove().draw();
-                            });
+                callback: function (result) {
+                    console.log(result);
+                    if (result) {
+                        $.ajax({
+                            method: "DELETE",
+                            url: `${url}/api/v1/items/${id}`,
+                            dataType: "json",
+                            contentType: 'application/json; charset=utf-8',
+                            headers: {
+                                "Authorization": "Bearer " + getToken()
+                            },
+                            success: function (data) {
+                                console.log(data);
+                                $row.fadeOut(4000, function () {
+                                    table.row($row).remove().draw();
+                                });
 
-                            bootbox.alert(data.success);
-                        },
-                        error: function (error) {
-                            bootbox.alert(data.error);
-                        }
-                    });
+                                bootbox.alert(data.success);
+                            },
+                            error: function (error) {
+                                console.log(error);
+                            }
+                        });
+
+                    }
 
                 }
+            });
 
-            }
-        });
-
-        // }
+        }
 
     })
 })
